+++ test/unit/lib/api/services.specs.js
@@ -11,16 +11,19 @@ test.describe('services api', () => {
     let operations
     let commandsMocks
     let baseMocks
+    let eventsMocks
 
     test.beforeEach(() => {
       baseMocks = new mocks.Base()
       commandsMocks = new mocks.Commands()
       operations = services.Operations(baseMocks.stubs.service, commandsMocks.stubs)
+      eventsMocks = new mocks.Events()
     })
 
     test.afterEach(() => {
       baseMocks.restore()
       commandsMocks.restore()
+      eventsMocks.restore()
     })
 
     test.describe('getServices handler', () => {
@@ -28,11 +31,27 @@ test.describe('services api', () => {
         const fooResult = 'foo result'
         commandsMocks.stubs.service.getFiltered.resolves(fooResult)
 
-        return operations.getServices.handler()
+        return operations.getServices.handler({query: {}})
           .then((result) => {
             return Promise.all([
               test.expect(result).to.equal(fooResult),
               test.expect(commandsMocks.stubs.service.getFiltered).to.have.been.calledWith()
+            ])
+          })
+      })
+
+      test.it('should pass received query data to get services command', () => {
+        const fooResult = 'foo result'
+        const fooQuery = {
+          type: 'foo-type'
+        }
+        commandsMocks.stubs.service.getFiltered.resolves(fooResult)
+
+        return operations.getServices.handler({query: fooQuery})
+          .then((result) => {
+            return Promise.all([
+              test.expect(result).to.equal(fooResult),
+              test.expect(commandsMocks.stubs.service.getFiltered).to.have.been.calledWith(fooQuery)
             ])
           })
       })
@@ -151,6 +170,13 @@ test.describe('services api', () => {
           })
       })
 
+      test.it('should emit a plugin event', () => {
+        return operations.addService.handler({}, fooBody, response, fooUserData)
+          .then(() => {
+            return test.expect(eventsMocks.stubs.plugin).to.have.been.calledWith('service', 'created', fooService)
+          })
+      })
+
       test.it('should resolve the promise with no value', () => {
         return operations.addService.handler({}, fooBody, response)
           .then((result) => {
@@ -249,6 +275,17 @@ test.describe('services api', () => {
         }, fooBody, response)
           .then(() => {
             return test.expect(response.header).to.have.been.calledWith('location', '/api/services/foo-service-id')
+          })
+      })
+
+      test.it('should emit a plugin event', () => {
+        return operations.updateService.handler({
+          path: {
+            id: fooId
+          }
+        }, fooBody, response)
+          .then(() => {
+            return test.expect(eventsMocks.stubs.plugin).to.have.been.calledWith('service', 'updated', fooService)
           })
       })